Added a global color system
